# docs/swagger.yaml
definitions:
  request.AddressAdd:
    properties:
      handle_status:
        description: 指派状态 0：所有，1：启用，2：停用
        example: 1
        type: integer
      num:
        description: 生成钱包数量
        type: integer
      use_tag:
        description: -2:eth钱包,-1:热钱包,0:未占用,userId:组织充币钱包
        example: 0
        type: integer
      user_id:
        description: 组织id
        example: 1
        type: integer
    type: object
  request.DetailedAdd:
    properties:
      amount:
        description: 金额
        type: number
      charge_address:
        description: 收款地址
        type: string
      create_at:
        description: 时间
        type: string
      from_address:
        description: 发款地址
        type: string
      id_code:
        description: 系统编号
        type: string
      txid:
        description: 交易hash
        type: string
    type: object
  request.DetailedEdit:
    properties:
      id:
        description: id
        type: integer
      is_bind:
        description: 1：解绑 2：绑定
        type: integer
      order_code:
        description: 商户订单编号
        type: string
    type: object
...
  request.MessageAdd:
    properties:
      beginTime:
        description: 开始时间
        example: "2020-09-15T14:41:46+08:00"
        type: string
      endTime:
        description: 结束时间
        example: "2020-09-23T14:41:50+08:00"
        type: string
      title:
        description: 系统公告
        example: 这是一个系统公告
        type: string
      users:
        description: 发给用户id
        example:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        items:
          type: integer
        type: array
    required:
    - beginTime
    - endTime
    - title
    - users
    type: object
  request.MessageEdit:
    properties:
      beginTime:
        description: 开始时间
        example: "2020-09-15T14:41:46+08:00"
        type: string
      endTime:
        description: 结束时间
        example: "2020-09-23T14:41:50+08:00"
        type: string
      id:
        type: integer
      title:
        description: 系统公告
        example: 这是一个系统公告
        type: string
      users:
        description: 发给用户id
        example:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        items:
          type: integer
        type: array
    required:
    - beginTime
    - endTime
    - title
    - users
    type: object
  request.OpenOrStopAddress:
    properties:
      hand_status:
        description: 是否启用
        example: 1
        type: integer
    type: object
  request.OrderAdd:
    properties:
      amount:
        description: 数量
        example: 1
        type: number
      orderCode:
        description: 商户订单编号
        example: asfasgdsasfgas
        type: string
      remark:
        description: 备注
        example: 备注
        type: string
      user_id:
        description: 用户id
        example: 1
        type: integer
    required:
    - amount
    - orderCode
    - remark
    - user_id
    type: object
  request.OrderEdit:
    properties:
      address:
        description: 收款地址
        type: string
      exprire_time:
        description: 过期时间
        type: string
      should_amount:
        description: 应收金额
        example: 1
        type: number
      status:
        description: 状态 -1：收款中，1：已完成，2：异常，3：退款等待中，4：退款中，5：退款失败，6：已退款，7：：已过期
        example: 3
        type: integer
    required:
    - address
    - should_amount
    - status
    type: object
...
  request.RateEdit:
    properties:
      pay_weight_add_or_reduce:
        description: 代发增加还是减少 0：增加 1：减少
        example: 0
        type: integer
      pay_weight_type:
        description: 代发加权类型：0：百分比，1：固定
        example: 0
        type: integer
      pay_weight_value:
        description: 代发加权值
        example: 1
        type: number
      receive_weight_add_or_reduce:
        description: 代收增加还是减少 0：增加 1：减少
        example: 0
        type: integer
      receive_weight_type:
        description: 代收加权类型：0：百分比，1：固定
        example: 0
        type: integer
      receive_weight_value:
        description: 代收加权值
        example: 1
        type: number
    required:
    - pay_weight_value
    - receive_weight_value
    type: object
...
  request.UpdateAmount:
    type: object
...
  response.AddressList:
    properties:
      create_time:
        description: 创建时间戳
        type: integer
      eth_amount:
        description: 以太币余额
        type: number
      id:
        description: ID
        type: integer
      real_name:
        description: 组织名称
        type: string
      status:
        description: 状态 1：已完成，2：执行中，3：结账中
        type: integer
      update_time:
        description: 更新时间戳
        type: integer
      usdt_amount:
        description: 泰达币余额
        type: number
      user_address:
        description: 地址
        type: string
      user_id:
        description: 商户id
        type: integer
      user_tag:
        description: 占用状态 0：停用，1：启用
        type: integer
    type: object
  response.AddressPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.AddressList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
  response.AdminSystemMessage:
    properties:
      begin_time:
        description: 开始时间
        type: string
      end_time:
        description: 结束时间
        type: string
      id:
        type: integer
      title:
        description: 标题
        type: string
    type: object
...
  response.AdminUserMessageList:
    properties:
      id:
        type: integer
      systemMessages:
        description: 系统公告
        items:
          $ref: '#/definitions/response.AdminSystemMessage'
        type: array
      user_name:
        description: 用户名
        type: string
    type: object
  response.AdminUserMessagePage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.AdminUserMessageList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
...
  response.DetailedList:
    properties:
      amount:
        description: 金额
        type: number
      charge_address:
        description: 收款地址
        type: string
      create_at:
        description: 时间
        type: string
      id:
        description: ID
        type: integer
      id_code:
        description: 系统编号
        type: string
      order_code:
        description: 订单编号
        type: string
      order_id:
        description: 商户订单id
        type: integer
      real_name:
        description: 组织名称
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态 1：未绑定 2：已绑定
        type: integer
      txid:
        description: 交易hash
        type: string
      user_id:
        description: 商户id
        type: integer
    type: object
  response.DetailedRecordPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/response.DetailedList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
...
  response.OrderPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.RealOrderList'
        type: array
      merchant_amount:
        description: 总商户总实收金额
        type: number
      per_page:
        description: 当前页码
        type: integer
      really_amount:
        description: 总实收金额
        type: number
      total:
        description: 总共多少页
        type: integer
      total_amount:
        description: 总订单金额
        type: number
      total_fee:
        description: 总手续费
        type: number
    type: object
...
  response.Rate:
    properties:
      collection:
        description: 代收加权
        type: number
      currency:
        description: 币别
        type: string
      id:
        description: ID
        type: integer
      original_rate:
        description: 原汇率
        type: number
      pay_weight_add_or_reduce:
        description: 代发增加还是减少 0：增加 1：减少
        type: integer
      pay_weight_type:
        description: 代发加权类型：0：百分比，1：固定
        type: integer
      pay_weight_value:
        description: 代发加权值
        type: number
      payment:
        description: 代发加权
        type: number
      receive_weight_add_or_reduce:
        description: 代收增加还是减少 0：增加 1：减少
        type: integer
      receive_weight_type:
        description: 代收加权类型：0：百分比，1：固定
        type: integer
      receive_weight_value:
        description: 代收加权值
        type: number
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  response.RealOrderList:
    properties:
      address:
        description: 收款地址
        type: string
      amount:
        description: 金额
        type: number
      create_at:
        description: 创建时间
        type: string
      detailed_record_id:
        description: 实收明细订单编号
        type: number
      fee:
        description: 手续费
        type: number
      finish_time:
        description: 完成时间
        type: string
      id:
        description: ID
        type: integer
      merchant_amount:
        description: 商户实收金额
        type: number
      order_code:
        description: 商户订单编号
        type: string
      real_name:
        description: 组织名称
        type: string
      receipt_amount:
        description: 实收金额
        type: number
      remark:
        description: 备注
        type: string
      return_amount:
        description: 退款金额
        type: number
      should_amount:
        description: 应收金额
        type: number
      status:
        description: 状态 -1：收款中，1：已完成，2：异常，3：退款等待中，4：退款中，5：退款失败，6：已退款，7：：已过期
        type: integer
      txid:
        description: 交易hash
        type: string
      user_id:
        description: 商户id
        type: integer
    type: object
...
  response.SystemMessageList:
    properties:
      begin_time:
        description: 开始时间
        type: string
      end_time:
        description: 结束时间
        type: string
      id:
        description: id
        type: integer
      status:
        description: 是否作废 0：否，1：是
        type: integer
      title:
        description: 标题
        type: string
      users:
        description: 用户名称
        items:
          $ref: '#/definitions/response.UserNames'
        type: array
    type: object
  response.SystemMessagePage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.SystemMessageList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
  response.UserNames:
    properties:
      id:
        description: 用户ID
        type: integer
      user_name:
        description: 用户名
        type: string
    type: object
info:
  contact: {}
  license: {}
  title: y2pay
paths:
...
  /addrList:
    get:
      parameters:
      - description: 状态 0：所有，1：已完成，2：执行中，3：结账中
        in: query
        name: status
        type: integer
      - description: 指派状态 0：所有，1：启用，2：停用
        in: query
        name: handleStatus
        type: integer
      - description: 钱包地址
        in: query
        name: address
        type: string
      - description: 商户组织id
        in: query
        name: userId
        required: true
        type: integer
      - description: 钱包占用类型 -2:eth钱包地址,-1：hot钱包地址,1:商户充币钱包地址
        in: query
        name: useTag
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AddressPage'
      summary: 钱包地址列表
      tags:
      - 地址管理
  /addrRestart:
    post:
      parameters:
      - description: 地址id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OpenOrStopAddress'
      produces:
      - application/json
      summary: 启用 停用地址
      tags:
      - 地址管理
...
  /createAddress:
    post:
      parameters:
      - description: 钱包地址
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AddressAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 为商户分配地址 生成热钱包地址 生成eth地址
      tags:
      - 地址管理
  /detail:
    get:
      parameters:
      - description: 状态 1：未绑定，2：已绑定
        in: query
        name: status
        type: integer
      - description: 系统编号
        in: query
        name: idCode
        type: string
      - description: 商户id
        in: query
        name: userId
        type: integer
      - description: 交易hash
        in: query
        name: txid
        type: string
      - description: 收款地址
        in: query
        name: address
        type: string
      - description: 起
        in: query
        name: from_date
        type: string
      - description: 至
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DetailedRecordPage'
      summary: 实收明细记录列表
      tags:
      - 实收明细记录
    post:
      parameters:
      - description: 实收记录
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DetailedAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 新增实收明细记录
      tags:
      - 实收明细记录
  /detail/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DetailedList'
      summary: 实收记录详情
      tags:
      - 实收明细记录
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 明细记录
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DetailedEdit'
      produces:
      - application/json
      summary: 绑定 | 解绑 订单
      tags:
      - 实收明细记录
...
  /order:
    get:
      parameters:
      - description: 状态 -1：收款中，1：已完成，2：异常，3：退款等待中，4：退款中，5：退款失败，6：已退款，7：已过期
        in: query
        name: status
        type: integer
      - description: 商户订单编号
        in: query
        name: orderCode
        type: string
      - description: 商户id
        in: query
        name: userId
        required: true
        type: integer
      - description: 交易哈希
        in: query
        name: txid
        type: string
      - description: 收款地址
        in: query
        name: chargeAddress
        type: string
      - description: 起
        in: query
        name: from_date
        type: string
      - description: 至
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderPage'
      summary: 订单列表
      tags:
      - 订单管理
    post:
      parameters:
      - description: 新增订单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OrderAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 新增订单
      tags:
      - 订单管理
  /order/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RealOrderList'
      summary: 订单详情
      tags:
      - 订单管理
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 实收订单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OrderEdit'
      produces:
      - application/json
      summary: 修改订单
      tags:
      - 订单管理
...
  /orderNotify:
    post:
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302": {}
      summary: 通知
      tags:
      - 订单管理
...
  /rate/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Rate'
      summary: 获取汇率详情
      tags:
      - 汇率管理
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 汇率
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RateEdit'
      produces:
      - application/json
      summary: 更新汇率
      tags:
      - 汇率管理
...
  /systemMessage:
    get:
      parameters:
      - description: 标题
        in: query
        name: title
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SystemMessagePage'
      summary: 获取系统公告列表
      tags:
      - 系统公告
    post:
      parameters:
      - description: 系统公告
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.MessageAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 添加系统公告
      tags:
      - 系统公告
  /systemMessage/{id}:
    delete:
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      summary: 删除公告
      tags:
      - 系统公告
    put:
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 公告
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.MessageEdit'
      produces:
      - application/json
      summary: 编辑公告
      tags:
      - 系统公告
  /systemMessageByUser:
    get:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AdminUserMessagePage'
      summary: 获取用户公告列表
      tags:
      - 系统公告
...
  /updateBalance:
    post:
      parameters:
      - description: 地址id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAmount'
      produces:
      - application/json
      summary: 更新余额
      tags:
      - 地址管理
swagger: "2.0"